+++ src/app/app.module.ts
@@ -2,8 +2,8 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { GalleryComponent } from './gallery_component/gallery.component';
-import { ImgComponent } from './img_component/img.component';
+import { GalleryComponent } from './gallery/gallery.component';
+import { ImgComponent } from './selected-qual-img/img.component';
 
 @NgModule({
   declarations: [
