+++ src/app/selected-qual-img/img.component.ts
