+++ src/app/selected-qual-img/style.css
