+++ tslint.json
@@ -1,5 +1,8 @@
 {
-  "extends": "tslint:recommended",
+  "extends": [
+    "tslint:recommended",
+    "tslint-config-airbnb"
+  ],
   "rules": {
     "array-type": false,
     "arrow-parens": false,
